Move React Comments, DataTableBase, Editor, and Protected Route into their own Folder
