+++ frontend/src/components/JobSiteList.js
@@ -3,7 +3,7 @@ import { Link, useNavigate } from 'react-router-dom';
 import { useApiRequest } from '../utils/useApiRequest';
 import { JOB_SITE_API_URL, formatDisplayDate  } from "../constants";
 
-import DataTableBase from './DataTableBase';
+import DataTableBase from './shared/DataTableBase';
 import {Button, Container, Row, Col} from 'reactstrap';
 
 const JobSiteList = () => {
