+++ frontend/src/components/shared/Editor.jsx
