+++ frontend/src/components/shared/ProtectedRoute.jsx
